+++ CHANGELOG.md
@@ -9,6 +9,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## Unreleased
 
+### Changed
+- resource `pullzone`: `websockets_max_connections` can be configured to increments of 100;
+
 ## 0.14.2 - 2026-05-19
 
 ### Fixed